refactor: merge Service and ReleaseService
